Create README.md
Hello everyone, I created a standard curriculum template for you to edit with html5 code and also using bootstrap.

I have been looking for a tool for a long time to be able to compile my information and apply it to a programmer vacancy.

I left it in two formats, one in PT-BR my native language and one in EN. I hope you use it and any improvement, we will add!

After finishing editing the curriculum, use the option to print and export as pdf on the web. Take care that it is important to select the Background graphics box to get the colors.